Make buttons on profile page visible on mobile
